# Opt CPU Backend - Vectorize across 8 element batches