--- a/backends/optimized/ceed-opt-basis.c
+++ b/backends/optimized/ceed-opt-basis.c
@@ -22,12 +22,13 @@
 // NOTRANSPOSE: V_ajc = T_jb U_abc
 // TRANSPOSE:   V_ajc = T_bj U_abc
 // If Add != 0, "=" is replaced by "+="
-static int CeedTensorContract_Opt(Ceed ceed,
+static int CeedTensorContract_Opt(Ceed ceed, CeedInt nelem,
                                   CeedInt A, CeedInt B, CeedInt C, CeedInt J,
-                                  const CeedScalar *restrict t, CeedTransposeMode tmode,
-                                  const CeedInt Add,
+                                  const CeedScalar *restrict t,
+                                  CeedTransposeMode tmode, const CeedInt Add,
                                   const CeedScalar *restrict u, CeedScalar *restrict v) {
   CeedInt tstride0 = B, tstride1 = 1;
+  const CeedInt blksize = 8;
   if (tmode == CEED_TRANSPOSE) {
     tstride0 = 1; tstride1 = J;
   }
@@ -38,15 +39,23 @@ static int CeedTensorContract_Opt(Ceed ceed,
     }
   }
 
-  for (CeedInt a=0; a<A; a++) {
-    for (CeedInt b=0; b<B; b++) {
-      for (CeedInt j=0; j<J; j++) {
-        CeedScalar tq = t[j*tstride0 + b*tstride1];
-        for (CeedInt c=0; c<C; c++) {
-          v[(a*J+j)*C+c] += tq * u[(a*B+b)*C+c];
+  if (nelem == 1) {
+    for (CeedInt a=0; a<A; a++)
+      for (CeedInt b=0; b<B; b++)
+        for (CeedInt j=0; j<J; j++) {
+          CeedScalar tq = t[j*tstride0 + b*tstride1];
+          for (CeedInt c=0; c<C; c++)
+            v[(a*J+j)*C+c] += tq * u[(a*B+b)*C+c];
         }
-      }
-    }
+  } else {
+    for (CeedInt a=0; a<A; a++)
+      for (CeedInt b=0; b<B; b++)
+        for (CeedInt j=0; j<J; j++) {
+          CeedScalar tq = t[j*tstride0 + b*tstride1];
+          for (CeedInt c=0; c<C; c+=blksize)
+            for (CeedInt e=0; e<blksize; e++)
+              v[((a*J+j)*C+c)+e] += tq * u[((a*B+b)*C+c)+e];
+        }
   }
   return 0;
 }
@@ -57,108 +66,116 @@ static int CeedBasisApply_Opt(CeedBasis basis, CeedInt nelem,
   int ierr;
   const CeedInt dim = basis->dim;
   const CeedInt ncomp = basis->ncomp;
-  const CeedInt nqpt = ncomp*CeedPowInt(basis->Q1d, dim);
+  const CeedInt nqpt = CeedPowInt(basis->Q1d, dim);
   const CeedInt add = (tmode == CEED_TRANSPOSE);
+  const CeedInt blksize = 8;
 
-  if (nelem != 1)
+  if ((nelem != 1) && (nelem != blksize))
     return CeedError(basis->ceed, 1,
-                     "This backend does not support BasisApply for multiple elements");
+                     "This backend does not support BasisApply for %d elements", nelem);
 
   if (tmode == CEED_TRANSPOSE) {
-    const CeedInt vsize = ncomp*CeedPowInt(basis->P1d, dim);
+    const CeedInt vsize = nelem*ncomp*CeedPowInt(basis->P1d, dim);
     for (CeedInt i = 0; i < vsize; i++)
-      v[i] = (CeedScalar) 0;
+      v[i] = (CeedScalar) 0.0;
   }
-  if (emode & CEED_EVAL_INTERP) {
+  if (emode == CEED_EVAL_INTERP) {
     CeedInt P = basis->P1d, Q = basis->Q1d;
     if (tmode == CEED_TRANSPOSE) {
       P = basis->Q1d; Q = basis->P1d;
     }
-    CeedInt pre = ncomp*CeedPowInt(P, dim-1), post = 1;
-    CeedScalar tmp[2][ncomp*Q*CeedPowInt(P>Q?P:Q, dim-1)];
+    CeedInt pre = ncomp*CeedPowInt(P, dim-1), post = nelem;
+    CeedScalar tmp[2][nelem*ncomp*Q*CeedPowInt(P>Q?P:Q, dim-1)];
     for (CeedInt d=0; d<dim; d++) {
-      ierr = CeedTensorContract_Opt(basis->ceed, pre, P, post, Q, basis->interp1d,
-                                    tmode, add&&(d==dim-1),
+      ierr = CeedTensorContract_Opt(basis->ceed, nelem, pre, P, post, Q,
+                                    basis->interp1d, tmode, add&&(d==dim-1),
                                     d==0?u:tmp[d%2], d==dim-1?v:tmp[(d+1)%2]);
       CeedChk(ierr);
       pre /= P;
       post *= Q;
     }
-    if (tmode == CEED_NOTRANSPOSE) {
-      v += nqpt;
-    } else {
-      u += nqpt;
-    }
-  }
-  if (emode & CEED_EVAL_GRAD) {
-    CeedInt P = basis->P1d, Q = basis->Q1d;
+  } else if (emode == CEED_EVAL_GRAD) {
     // In CEED_NOTRANSPOSE mode:
-    // u is (P^dim x nc), column-major layout (nc = ncomp)
-    // v is (Q^dim x nc x dim), column-major layout (nc = ncomp)
+    // u is (P^dim x nc) x nelem, column-major layout (nc = ncomp)
+    // v is (Q^dim x nc x dim) x nelem, column-major layout (nc = ncomp)
     // In CEED_TRANSPOSE mode, the sizes of u and v are switched.
-    if (tmode == CEED_TRANSPOSE) {
-      P = basis->Q1d, Q = basis->P1d;
-    }
     if (dim == 1) {
-      ierr = CeedTensorContract_Opt(basis->ceed, ncomp, P, 1, Q, basis->grad1d,tmode,
-                                    add, u, v); CeedChk(ierr);
+      CeedInt P = basis->P1d, Q = basis->Q1d;
+      if (tmode == CEED_TRANSPOSE) {
+        P = basis->Q1d, Q = basis->P1d;
+      }
+      ierr = CeedTensorContract_Opt(basis->ceed, nelem, ncomp, P, nelem, Q,
+                                    basis->grad1d, tmode, add, u, v);
+      CeedChk(ierr);
     } else {
+      CeedInt P = basis->P1d, Q = basis->Q1d;
       if (tmode == CEED_TRANSPOSE) {
         P = basis->Q1d, Q = basis->Q1d;
       }
       CeedBasis_Opt *impl = basis->data;
-      CeedScalar interp[ncomp*Q*CeedPowInt(P>Q?P:Q, dim-1)];
-      CeedInt pre = ncomp*CeedPowInt(P, dim-1), post = 1;
-      CeedScalar tmp[2][ncomp*Q*CeedPowInt(P>Q?P:Q, dim-1)];
+      CeedScalar interp[nelem*ncomp*Q*CeedPowInt(P>Q?P:Q, dim-1)];
+      CeedInt pre = ncomp*CeedPowInt(P, dim-1), post = nelem;
+      CeedScalar tmp[2][nelem*ncomp*Q*CeedPowInt(P>Q?P:Q, dim-1)];
+      // Interpolate to quadrature points (NoTranspose)
+      //  or Grad to quadrature points (Transpose)
       for (CeedInt d=0; d<dim; d++) {
-        ierr = CeedTensorContract_Opt(basis->ceed, pre, P, post, Q,
-                                      tmode==CEED_NOTRANSPOSE?basis->interp1d:impl->colograd1d,
+        ierr = CeedTensorContract_Opt(basis->ceed, nelem, pre, P, post, Q,
+                                      tmode==CEED_NOTRANSPOSE
+                                        ? basis->interp1d
+                                        : impl->colograd1d,
                                       tmode, add&&(d>0),
-                                      tmode==CEED_NOTRANSPOSE?(d==0?u:tmp[d%2]):u,
-                                      tmode==CEED_NOTRANSPOSE?(d==dim-1?interp:tmp[(d+1)%2]):interp);
+                                      tmode==CEED_NOTRANSPOSE
+                                        ? (d==0?u:tmp[d%2])
+                                        : u + d*nqpt*ncomp*nelem,
+                                      tmode==CEED_NOTRANSPOSE
+                                        ? (d==dim-1?interp:tmp[(d+1)%2])
+                                        : interp);
         CeedChk(ierr);
         pre /= P;
         post *= Q;
-        if (tmode == CEED_TRANSPOSE) {
-          u += nqpt;
-        }
       }
+      // Grad to quadrature points (NoTranspose)
+      //  or Interpolate to dofs (Transpose)
       P = basis->Q1d, Q = basis->Q1d;
       if (tmode == CEED_TRANSPOSE) {
         P = basis->Q1d, Q = basis->P1d;
       }
-      pre = ncomp*CeedPowInt(P, dim-1), post = 1;
+      pre = ncomp*CeedPowInt(P, dim-1), post = nelem;
       for (CeedInt d=0; d<dim; d++) {
-        ierr = CeedTensorContract_Opt(basis->ceed, pre, P, post, Q,
-                                      tmode==CEED_NOTRANSPOSE?impl->colograd1d:basis->interp1d,
+        ierr = CeedTensorContract_Opt(basis->ceed, nelem, pre, P, post, Q,
+                                      tmode==CEED_NOTRANSPOSE
+                                        ? impl->colograd1d
+                                        : basis->interp1d,
                                       tmode, add&&(d==dim-1),
-                                      tmode==CEED_NOTRANSPOSE?interp:(d==0?interp:tmp[d%2]),
-                                      tmode==CEED_NOTRANSPOSE?v:(d==dim-1?v:tmp[(d+1)%2]));
+                                      tmode==CEED_NOTRANSPOSE
+                                        ? interp
+                                        : (d==0?interp:tmp[d%2]),
+                                      tmode==CEED_NOTRANSPOSE
+                                        ? v + d*nqpt*ncomp*nelem
+                                        : (d==dim-1?v:tmp[(d+1)%2]));
         CeedChk(ierr);
         pre /= P;
         post *= Q;
-        if (tmode == CEED_NOTRANSPOSE) {
-          v += nqpt;
-        }
       }
     }
-  }
-  if (emode & CEED_EVAL_WEIGHT) {
+  } else if (emode == CEED_EVAL_WEIGHT) {
     if (tmode == CEED_TRANSPOSE)
       return CeedError(basis->ceed, 1,
                        "CEED_EVAL_WEIGHT incompatible with CEED_TRANSPOSE");
     CeedInt Q = basis->Q1d;
     for (CeedInt d=0; d<dim; d++) {
       CeedInt pre = CeedPowInt(Q, dim-d-1), post = CeedPowInt(Q, d);
-      for (CeedInt i=0; i<pre; i++) {
-        for (CeedInt j=0; j<Q; j++) {
+      for (CeedInt i=0; i<pre; i++)
+        for (CeedInt j=0; j<Q; j++)
           for (CeedInt k=0; k<post; k++) {
-            v[(i*Q + j)*post + k] = basis->qweight1d[j]
-                                    * (d == 0 ? 1 : v[(i*Q + j)*post + k]);
-          }
+            CeedScalar w = basis->qweight1d[j]
+                           * (d == 0 ? 1 : v[((i*Q + j)*post + k)*nelem]);
+            for (CeedInt e=0; e<nelem; e++)
+              v[((i*Q + j)*post + k)*nelem + e] = w;
         }
-      }
     }
+  } else if (emode == CEED_EVAL_DIV) {
+  } else if (emode == CEED_EVAL_CURL) {
   }
   return 0;
 }