… backends/optimized/ceed-opt-operator.c
… output/t08-elemrestriction-f.out
… output/t08-elemrestriction.out