I have not done speed tests yet, but the new vectorization is in place.