@jeremylt 2018-07-28
# Crude initial timing results:

## Current Master with `/cpu/self/opt`

```
$ time ./build/petsc-bp1 -local 1000000
Process decomposition: 1 1 1
Local elements: 1000000 = 100 100 100
Owned dofs: 1030301 = 101 101 101
KSP cg CONVERGED_RTOL iterations 23 rnorm 8.518818e-08
Pointwise error (max) 2.696996e-04

real	0m8.216s
user	0m7.994s
sys	0m0.217s

```

## New Branch with `/cpu/self/opt`

```
$ time ./build/petsc-bp1 -local 500000
Process decomposition: 1 1 1
Local elements: 1000000 = 100 100 100
Owned dofs: 1030301 = 101 101 101
KSP cg CONVERGED_RTOL iterations 23 rnorm 8.518818e-08
Pointwise error (max) 2.696996e-04

real	0m8.038s
user	0m7.803s
sys	0m0.227s
```

This feels like a rather small improvement, though I could be running too small of a problem or my vectorization could stand to be improved or both.