@jeremylt 2018-08-05
This PR will be updated following PR #110 and PR #106 to improve handling of batches of elements.