@jeremylt 2018-08-08
# New timing results:

## Current Master with `/cpu/self/ref`

```
$ time ./build/mfem-bp3 -ceed /cpu/self/ref -o 6
Options used:
   -ceed /cpu/self/ref
   --mesh /home/jeremy/Dev/mfem/data/star.mesh
   --order 6
   --visualization
   --no-test
Number of finite element unknowns: 46561
   Iteration :   0  (B r, r) = 1697.41 ...
   Iteration : 662  (B r, r) = 1.57191e-09
Average reduction factor = 0.97929
L2 projection error: 9.83612e-06

real	0m4.173s
user	0m4.164s
sys	0m0.003s
```

## Current Master with `/cpu/self/opt`

```
$ time ./build/mfem-bp3 -ceed /cpu/self/opt -o 6
Options used:
   -ceed /cpu/self/opt
   --mesh /home/jeremy/Dev/mfem/data/star.mesh
   --order 6
   --visualization
   --no-test
Number of finite element unknowns: 46561
   Iteration :   0  (B r, r) = 1697.41 ...
   Iteration : 662  (B r, r) = 1.57191e-09
Average reduction factor = 0.97929
L2 projection error: 9.83612e-06

real	0m4.202s
user	0m4.192s
sys	0m0.007s
```

## New Branch with `/cpu/self/opt`

```
$ time ./build/mfem-bp3 -ceed /cpu/self/opt -o 6
Options used:
   -ceed /cpu/self/opt
   --mesh /home/jeremy/Dev/mfem/data/star.mesh
   --order 6
   --visualization
   --no-test
Number of finite element unknowns: 46561
   Iteration :   0  (B r, r) = 1697.41 ...
   Iteration : 662  (B r, r) = 1.57191e-09
Average reduction factor = 0.97929
L2 projection error: 9.83612e-06

real	0m2.736s
user	0m2.719s
sys	0m0.013s
```

Much better!