@jeremylt 2018-08-08
At some point along the line we stopped testing `/cpu/self/tmpl`. Was that intentional?